+++ src/hooks/useAllowance.ts
@@ -1,35 +1,35 @@
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useState, useRef } from "react";
 
 import BigNumber from "bignumber.js";
 import { useWallet } from "use-wallet";
 import { provider } from "web3-core";
-import { getAllowance, shouldUpdateVal } from "utils";
+import { getAllowance } from "utils";
+import useUbiq from "hooks/useUbiq";
 
 const useAllowance = (tokenAddress?: string, spenderAddress?: string) => {
   const [allowance, setAllowance] = useState<BigNumber>();
   const { account, ethereum }: { account: string | null; ethereum?: provider } = useWallet();
 
-  const fetchAllowance = useCallback(
-    async (userAddress: string, provider: provider) => {
-      if (!spenderAddress || !tokenAddress) {
-        return;
-      }
-      const newAllowance = new BigNumber(await getAllowance(userAddress, spenderAddress, tokenAddress, provider));
+  const { BlockNum } = useUbiq();
+  const alreadyAllowed = useRef(false);
 
-      if (shouldUpdateVal(allowance, newAllowance, "BigNumber")) {
-        setAllowance(newAllowance);
-      }
-    },
-    [allowance, setAllowance, spenderAddress, tokenAddress]
-  );
+  const fetchAllowance = useCallback(async () => {
+    if (!spenderAddress || !tokenAddress || !ethereum || !account) {
+      return;
+    }
+    if (alreadyAllowed.current) return;
+
+    const newAllowance = new BigNumber(await getAllowance(account, spenderAddress, tokenAddress, ethereum));
+
+    if (newAllowance.toNumber() > 0) {
+      alreadyAllowed.current = true;
+    }
+    setAllowance(newAllowance);
+  }, [setAllowance, spenderAddress, tokenAddress, account, ethereum]);
 
   useEffect(() => {
-    if (account && ethereum && spenderAddress && tokenAddress) {
-      fetchAllowance(account, ethereum);
-    }
-    let refreshInterval = setInterval(fetchAllowance, 10000);
-    return () => clearInterval(refreshInterval);
-  }, [account, ethereum, spenderAddress, tokenAddress, fetchAllowance]);
+    fetchAllowance();
+  }, [BlockNum, fetchAllowance]);
 
   return allowance;
 };